backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "cryptography>=44.0.0",
    "fastapi>=0.135.1",
    "httpx>=0.28.0",
    "pydantic-settings>=2.7.0",
    "pyjwt>=2.10.0",
    "sqlmodel>=0.0.37",
    "uvicorn>=0.42.0",
]
